あらら

+++ helth_survey/Result+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  Result+CoreDataProperties.swift
 //  helth_survey
 //
-//  Created by 杉山新 on 2024/07/17.
+//  Created by 杉山新 on 2024/07/29.
 //
 //
 
@@ -19,7 +19,7 @@ extension Result {
     @NSManaged public var evaluationData: Date?
     @NSManaged public var resultID: UUID?
     @NSManaged public var studentsID: UUID?
-    @NSManaged public var totalScore: Int16
+    @NSManaged public var totalScore: Int32
     @NSManaged public var student: Student?
 
 }
